+++ calendar-ui/src/lib/utils/timeAgo.ts
@@ -1,0 +1,34 @@
+export function timeAgo(date: Date) {
+  const seconds = Math.floor((Date.now() - date.getTime()) / 1000);
+  const intervals: [number, string][] = [
+    [60, 'second'],
+    [60, 'minute'],
+    [24, 'hour'],
+    [7, 'day'],
+    [4.34524, 'week'],
+    [12, 'month'],
+    [Number.POSITIVE_INFINITY, 'year'],
+  ];
+
+  let counter = seconds;
+  for (let i = 0; i < intervals.length; i++) {
+    const [limit, name] = intervals[i];
+    if (counter < limit) {
+      const value = Math.floor(counter) || 0;
+      return `${value} ${name}${value !== 1 ? 's' : ''}`;
+    }
+    counter = Math.floor(counter / limit);
+  }
+  return '0 seconds';
+}
+
+export function timeAgoShort(dateStr: string) {
+  const d = new Date(dateStr);
+  const seconds = Math.floor((Date.now() - d.getTime()) / 1000);
+  if (seconds < 60) return `${seconds}s`;
+  const mins = Math.floor(seconds / 60);
+  if (mins < 60) return `${mins}m`;
+  const hrs = Math.floor(mins / 60);
+  if (hrs < 24) return `${hrs}h`;
+  return d.toLocaleDateString();
+}
