+++ calendar-ui/src/pages/EntryDetails.tsx
@@ -16,6 +16,7 @@ import {
 import { useLocation } from 'react-router-dom';
 import io from 'socket.io-client';
 import { fetchActivity } from '../api/activitiesApi';
+import { timeAgo } from '../lib/utils/timeAgo';
 import type { ActivityResponse } from '@corpcal/shared/api/types';
 import {
   DocumentText16Regular,
@@ -85,30 +86,6 @@ const useStyles = makeStyles({
     fontWeight: 500,
   },
 });
-
-function timeAgo(date: Date) {
-  const seconds = Math.floor((Date.now() - date.getTime()) / 1000);
-  const intervals: [number, string][] = [
-    [60, 'second'],
-    [60, 'minute'],
-    [24, 'hour'],
-    [7, 'day'],
-    [4.34524, 'week'],
-    [12, 'month'],
-    [Number.POSITIVE_INFINITY, 'year'],
-  ];
-
-  let counter = seconds;
-  for (let i = 0; i < intervals.length; i++) {
-    const [limit, name] = intervals[i];
-    if (counter < limit) {
-      const value = Math.floor(counter) || 0;
-      return `${value} ${name}${value !== 1 ? 's' : ''}`;
-    }
-    counter = Math.floor(counter / limit);
-  }
-  return '0 seconds';
-}
 
 export const EntryDetails = () => {
   const location = useLocation();
